♻️ [refactor] #133 : Exercise API의 Dto 로직 분리
