--- server/src/exercises/dto/every-date.dto.ts
@@ -1,9 +1,0 @@
-export class EveryDateDto {
-  dateList: string[] = [];
-
-  constructor(dateObject: { date: string }[]) {
-    dateObject.map((item) => {
-      this.dateList.push(item.date);
-    });
-  }
-}
